feat: code cleanup & settings persistance
