+++ app/reducers/index.ts
@@ -3,7 +3,7 @@ import browserReducer from './browser';
 import engineReducer, { controllerReducers } from '../core/redux/slices/engine';
 import privacyReducer from './privacy';
 import modalsReducer from './modals';
-import settingsReducer from './settings';
+import settingsReducer from '../core/redux/slices/settings';
 import alertReducer from './alert';
 import transactionReducer from './transaction';
 import userReducer from './user';
